RAC-450: Handle confirm when hitting enter on the Delete Attribute modal

+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/attribute/form/delete/DeleteModal.tsx
@@ -6,11 +6,13 @@ import {
   Field,
   getColor,
   Helper,
+  Key,
   Link,
   Modal,
   SectionTitle,
   TextInput,
   Title,
+  useShortcut,
 } from 'akeneo-design-system';
 import {NotificationLevel, useNotify, useTranslate, useRoute} from '@akeneo-pim-community/legacy-bridge';
 import {useIsMounted} from '@akeneo-pim-community/shared';
@@ -61,28 +63,31 @@ const DeleteModal = ({onCancel, onSuccess, attributeCode}: DeleteModalProps) => 
   const [attributeCodeConfirm, setAttributeCodeConfirm] = useState<string>('');
   const isValid = attributeCodeConfirm === attributeCode;
 
-  const handleConfirm = () => {
-    fetch(removeRoute, {
-      method: 'DELETE',
-      headers: new Headers({
-        'X-Requested-With': 'XMLHttpRequest',
-      }),
-    })
-      .then(async (response: Response) => {
-        if (response.ok) {
-          notify(NotificationLevel.SUCCESS, translate('pim_enrich.entity.attribute.flash.delete.success'));
-          onSuccess();
-        } else {
-          notify(
-            NotificationLevel.ERROR,
-            (await response.json()).message ?? translate('pim_enrich.entity.attribute.flash.delete.fail')
-          );
-        }
-      })
-      .catch(() => {
-        notify(NotificationLevel.ERROR, translate('pim_enrich.entity.attribute.flash.delete.fail'));
+  const handleConfirm = async () => {
+    if (!isValid) return;
+
+    try {
+      setAttributeCodeConfirm('');
+      const response = await fetch(removeRoute, {
+        method: 'DELETE',
+        headers: new Headers({
+          'X-Requested-With': 'XMLHttpRequest',
+        }),
       });
+
+      if (response.ok) {
+        notify(NotificationLevel.SUCCESS, translate('pim_enrich.entity.attribute.flash.delete.success'));
+        onSuccess();
+      } else {
+        const {message} = await response.json();
+        notify(NotificationLevel.ERROR, message ?? translate('pim_enrich.entity.attribute.flash.delete.fail'));
+      }
+    } catch (error) {
+      notify(NotificationLevel.ERROR, translate('pim_enrich.entity.attribute.flash.delete.fail'));
+    }
   };
+
+  useShortcut(Key.Enter, handleConfirm);
 
   const productText =
     0 < productCount
